requirements.txt
# Core Frameworks
torch>=2.0.0
transformers>=4.30.0
tokenizers>=0.13.3
numpy>=1.24.3

# Linguistic Databases & Analyzers
stanza>=1.5.0
nltk>=3.8.1

# Dataset & Evaluation
datasets>=2.12.0
evaluate>=0.4.0
sacrebleu>=2.3.1
rouge_score>=0.1.2

# Utilities
pyyaml>=6.0
tqdm>=4.65.0